refactor: change structure of the project and any rename
