+++ src/utilits/getCurrentSettings.ts
@@ -1,5 +1,13 @@
 import { carData } from '../car/car_data';
 import { ICarData, ISettingsMyCar } from '../types';
+import { defaultSettings } from '../constants/constants';
+import { setUserSettings } from '../helpers/authentication';
+
+export function mySetting(): ISettingsMyCar {
+  return localStorage.getItem('settingsCar')
+    ? JSON.parse(localStorage.getItem('settingsCar') as string)
+    : setUserSettings(defaultSettings);
+}
 
 export function getSettingsFromLocal(): ISettingsMyCar | undefined {
   if (localStorage.getItem('settingsCar')) {
